feat: moved all references and tests from base to inner_product for Curve25519Scalar, RistrettoPoint, and InnerProductProof
Format update
Fixed issue with CI
reexport needed to be made pub, not the module
Updated formatting
Fixed CI issues
Fixed according to reviews
Corrected imports
Formatting
Copied tests to inner_product
CI changes
Updated according to comments
refactor: Moved all references and tests from base to inner_product for Curve25518Scalar, RistrettoPoint, InnerProductProof
Fixed issues with tests
Fixed according to reviews
