+++ crates/proof-of-sql/src/base/arrow/scalar_and_i256_conversions.rs
@@ -58,11 +58,12 @@ mod tests {
     use super::{convert_i256_to_scalar, convert_scalar_to_i256};
     use crate::base::{
         database::scalar_and_i256_conversions::{MAX_SUPPORTED_I256, MIN_SUPPORTED_I256},
-        scalar::{test_scalar::TestScalar, Curve25519Scalar, Scalar},
+        scalar::{test_scalar::TestScalar, Scalar},
     };
     use arrow::datatypes::i256;
     use num_traits::Zero;
     use rand::RngCore;
+
     /// Generate a random i256 within a supported range. Values generated by this function will
     /// fit into the i256 but will not exceed 252 bits of width.
     fn random_i256<R: RngCore + ?Sized>(rng: &mut R) -> i256 {
@@ -89,12 +90,6 @@ mod tests {
         // Must fit inside 252 bits and so requires fallible
         fn try_from(value: i256) -> Result<Self, ()> {
             convert_i256_to_scalar(&value).ok_or(())
-        }
-    }
-
-    impl From<Curve25519Scalar> for i256 {
-        fn from(value: Curve25519Scalar) -> Self {
-            convert_scalar_to_i256(&value)
         }
     }
