+++ crates/proof-of-sql/src/base/slice_ops/slice_cast_test.rs
@@ -1,30 +1,10 @@
 use super::*;
-use crate::base::scalar::{test_scalar::TestScalar, Curve25519Scalar};
-use curve25519_dalek::scalar::Scalar;
-
+use crate::base::scalar::test_scalar::TestScalar;
 #[test]
 fn test_slice_map_to_vec() {
     let a: Vec<u32> = vec![1, 2, 3, 4];
     let b: Vec<u64> = vec![1, 2, 3, 4];
     let a: Vec<u64> = slice_cast_with(&a, |&x| u64::from(x));
-    assert_eq!(a, b);
-}
-
-/// add tests for [`slice_cast_with`]
-#[test]
-fn test_slice_cast_with_from_curve25519_scalar_to_dalek_scalar() {
-    let a: Vec<Curve25519Scalar> = vec![Curve25519Scalar::from(1u64), Curve25519Scalar::from(2u64)];
-    let b: Vec<Scalar> = vec![Scalar::from(1u64), Scalar::from(2u64)];
-    let a: Vec<Scalar> = slice_cast_with(&a, core::convert::Into::into);
-    assert_eq!(a, b);
-}
-
-/// add tests for [`slice_cast`]
-#[test]
-fn test_slice_cast_from_curve25519_scalar_to_dalek_scalar() {
-    let a: Vec<Curve25519Scalar> = vec![Curve25519Scalar::from(1u64), Curve25519Scalar::from(2u64)];
-    let b: Vec<Scalar> = vec![Scalar::from(1u64), Scalar::from(2u64)];
-    let a: Vec<Scalar> = slice_cast(&a);
     assert_eq!(a, b);
 }
