+++ crates/proof-of-sql/src/proof_primitive/inner_product/mod.rs
@@ -1,0 +1,9 @@
+/// TODO: add doc
+pub mod curve_25519_scalar;
+#[cfg(test)]
+mod curve_25519_tests;
+/// TODO: add doc
+#[cfg(feature = "blitzar")]
+pub mod inner_product_proof;
+/// TODO: add doc
+pub mod ristretto_point;
