+++ crates/proof-of-sql/src/sql/proof/verification_builder_test.rs
@@ -1,6 +1,6 @@
 use super::{SumcheckMleEvaluations, VerificationBuilderImpl};
 use crate::{
-    base::scalar::Curve25519Scalar,
+    proof_primitive::inner_product::curve_25519_scalar::Curve25519Scalar,
     sql::proof::{SumcheckSubpolynomialType, VerificationBuilder},
 };
 use alloc::collections::VecDeque;
