+++ crates/proof-of-sql/src/sql/proof_exprs/equals_expr_test.rs
@@ -5,8 +5,9 @@ use crate::{
             owned_table_utility::*, table_utility::*, Column, OwnedTable, OwnedTableTestAccessor,
             Table, TableRef, TableTestAccessor,
         },
-        scalar::{Curve25519Scalar, Scalar},
+        scalar::Scalar,
     },
+    proof_primitive::inner_product::curve_25519_scalar::Curve25519Scalar,
     sql::{
         proof::{exercise_verification, VerifiableQueryResult},
         proof_exprs::{test_utility::*, DynProofExpr, ProofExpr},
