+++ crates/proof-of-sql/src/sql/proof_exprs/multiply_expr_test.rs
@@ -5,8 +5,9 @@ use crate::{
             owned_table_utility::*, table_utility::*, Column, OwnedTableTestAccessor, TableRef,
             TableTestAccessor,
         },
-        scalar::{test_scalar::TestScalar, Curve25519Scalar},
+        scalar::test_scalar::TestScalar,
     },
+    proof_primitive::inner_product::curve_25519_scalar::Curve25519Scalar,
     sql::{
         parse::ConversionError,
         proof::{exercise_verification, QueryError, VerifiableQueryResult},
