+++ crates/proof-of-sql/src/sql/proof_gadgets/monotonic_test.rs
@@ -113,8 +113,8 @@ mod tests {
         base::{
             database::{table_utility::*, ColumnType, TableTestAccessor},
             math::decimal::Precision,
-            scalar::Curve25519Scalar,
         },
+        proof_primitive::inner_product::curve_25519_scalar::Curve25519Scalar,
         sql::proof::{QueryError, VerifiableQueryResult},
     };
     use blitzar::proof::InnerProductProof;
