+++ crates/proof-of-sql/tests/decimal_integration_tests.rs
@@ -4,8 +4,12 @@
 #[cfg(feature = "blitzar")]
 use blitzar::proof::InnerProductProof;
 #[cfg(feature = "blitzar")]
-use proof_of_sql::base::{database::owned_table_utility::*, scalar::Curve25519Scalar as S};
-use proof_of_sql::{base::database::TableRef, sql::postprocessing::apply_postprocessing_steps};
+use proof_of_sql::base::database::owned_table_utility::*;
+use proof_of_sql::{
+    base::database::TableRef,
+    proof_primitive::inner_product::curve_25519_scalar::Curve25519Scalar as S,
+    sql::postprocessing::apply_postprocessing_steps,
+};
 
 #[cfg(feature = "blitzar")]
 fn run_query(
@@ -57,7 +61,10 @@ fn run_query(
 #[cfg(feature = "blitzar")]
 mod decimal_query_tests {
     use crate::run_query;
-    use proof_of_sql::base::scalar::{Curve25519Scalar as S, Scalar};
+    use proof_of_sql::{
+        base::scalar::Scalar,
+        proof_primitive::inner_product::curve_25519_scalar::Curve25519Scalar as S,
+    };
 
     #[test]
     fn we_can_query_decimals_exactly_matching_db_data() {
